Total clicks (#348)
**Context:**

The computation of grouped (total) click probabilities is necessary for running validation tests in the context of Gaussian boson sampling with threshold detectors. Drummond et al. proposed an efficient method to compute these probabilities.

**Description of the Change:**

The computation of grouped click probabilities, following Drummond's simulation method, is implemented using numba.

**Benefits:**

The function, which is not implemented in thewalrus, gives a simple way to compute total click probability distributions.

**Possible Drawbacks:**

In some cases the number of samples needed for convergence may be large, thus increasing the time of computation.

---------

Co-authored-by: Sebastián Duque Mesa <675763+sduquemesa@users.noreply.github.com>
Co-authored-by: Nicolas Quesada <991946+nquesada@users.noreply.github.com>
Co-authored-by: Nicolas Quesada <nquesada@pop-os.localdomain>